edit language, units and windSpeed setting

+++ app/src/main/java/com/example/weatherapp/ui/favorite/view/FavoriteFragment.kt
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
@@ -36,6 +37,10 @@ class FavoriteFragment : Fragment() , OnClickFavorite{
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+    }
+
+    override fun onResume() {
+        super.onResume()
         (activity as AppCompatActivity?)!!.supportActionBar!!.hide()
     }
     override fun onStop() {
@@ -114,7 +119,15 @@ class FavoriteFragment : Fragment() , OnClickFavorite{
     }
 
     override fun onClickDelete(favoriteLocation: FavoriteLocation) {
-        viewModel.deleteFavFromDB(favoriteLocation)
+        AlertDialog.Builder(requireContext())
+            .setTitle(getString(R.string.delete_confirm))
+            .setMessage(getString(R.string.delete_qustion))
+            .setPositiveButton(android.R.string.ok) { _, _ ->
+                viewModel.deleteFavFromDB(favoriteLocation)
+            }
+            .setNegativeButton(android.R.string.cancel, null)
+            .show()
+
     }
 
 }
